update traverse
Instead of printing in the traverse function, just construct a list
of file objects that can later have formatting and sorting applied
to it.

Right now traverse returns a list of `java.io/file`s, but eventually
it will return a list of `Map`s with keys `:file-obj`, `:parent`,
and `:depth`.